AI now detects the player within it's field of view
The player needed a rigidbody for it to be detected by the AI, so that was added. That means the fps controller is going to have to be improved to support steps and ramps and stuff. The AI has a large sphere collider around it, and checks within a section for the player