Big update to detection systems
Didn't push for 20 days because i'm a bit tired of fixing everything the way we usually have to.
I started the audio detection system, if an object with the collision script collides with something it will make a pulse within a decent area, i have already tested that the AI can work with this system. Player sound needs to be implemented.
The new visual detection system was started and heavily worked on, biggest hurdle is currently that the raycasts move and rotate unexpectedly.
I created a script to generate cone meshes, the visual viewcone thing isn't perfected as it can be projected through walls.
I made a post processing shader to try and achieve a better result with the outline stuff and it kinda worked, just breaks on specific shapes or angles, and i need to look into how to use post processing more