Audio Detection System Update
Player now makes sound in different rates and ranges depending on if they are walking or crouching.
AI will go to collisions it hears or if it detects the player after hearing them for long enough.
Updated the rigidbody character controller to support crouching, currently launches the player off stairs and this is due to being positioned too close to the edge of a step and Unity physics corrects this by shooting the player out of the collider they are in.
There is another bug where the AI's detection will not decrease sometimes, that needs to be fixed asap